\name{Zuschlagskalkulation-package}
\alias{Zuschlagskalkulation-package}
\docType{package}
\title{Berechnet die Selbstkosten anhand einer Zuschlagskalkulation}
\description{
Berechnet die Selbstkosten anhand einer Zuschlagskalkulation mit einer Kostenstelle fuer kleine Handwerksbetriebe oder mehreren Kostenstellen fuer Industiebetriebe. Die Funktion *.handwerk ist vereinfacht und nutzt nur einen allgemeinen Gemeinkostenzuschlag. Die Funktion *.industrie ist ein wenig komplexer und benutzt mehrere Gemeinkostenzuschlaege. In beiden Funktionen werden Gemeinkostenzuschlaege in Prozent angegeben. Dieses Paket ist noch im Entwickungsstadium und soll noch um weitere Funktionen erweitert werden.  
}
\details{
\tabular{ll}{
Package: \tab Zuschlagskalkulation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-14\cr
License: \tab GNU General Public License\cr
}}
\author{
Christopher M. Stadin \cr
Maintainer: Bei Fragen E-Mail an <chrufru@gmail.com>
}
\references{
Gummersbach, Alfons -Produktionsmanagement ISBN:
}
\keyword{package}
\seealso{}
\examples{}
